refactor: move is...CapableAccount() utils to @core namespace
